+++ app/app_mobile_and.c
@@ -1,6 +1,5 @@
-// +build !ci
-
-// +build android
+//go:build !ci && android
+// +build !ci,android
 
 #include <android/log.h>
 #include <jni.h>