+++ theme/theme_other.go
@@ -1,3 +1,4 @@
+//go:build !hints
 // +build !hints
 
 package theme